cleanup routes and controllers
